params create ft job
